fix: insta 관련 함수 및 변수 insta.py로 이동, patch -> put 으로 수정
