+++ main.py
@@ -10,12 +10,10 @@ from starlette.responses import RedirectResponse
 from sqlalchemy.orm import Session
 
 from starlette.middleware.cors import CORSMiddleware
-import crud
 from utils import check_db_connected
-import models, schemas
+import models, schemas, crud, insta
 from database import SessionLocal, engine
 from voice_alteration import voice_alteration
-import crud
 
 
 models.Base.metadata.create_all(bind=engine)
@@ -80,24 +78,10 @@ def main():
     return RedirectResponse(url="/docs/")
 
 
-# ----------------------------------------------------
-# 인스타 로그인 및 프로필 정보 가져오기 및 유저 관련 API
-app_id = os.getenv('INSTA_APP_ID')
-secret_id = os.getenv('INSTA_APP_SECRET_ID')
-redirect_url = "https://localhost:8000/api/v1/insta/redirection"
-authorize_url = f"https://api.instagram.com/oauth/authorize?client_id=\
-    {app_id}&redirect_uri={redirect_url}&scope=user_profile,user_media&response_type=code"
-get_short_token_url = "https://api.instagram.com/oauth/access_token"
-get_user_name_url = "https://graph.instagram.com/me?fields=username&access_token="
-get_user_info_url = "https://i.instagram.com/api/v1/users/web_profile_info/?username="
-get_long_token_url = f"https://graph.instagram.com/access_token?grant_type=ig_exchange_token&client_secret={secret_id}&access_token="
-refresh_token_url = "https://graph.instagram.com/refresh_access_token?grant_type=ig_refresh_token&access_token="
-
-
 # 앱 접속 시 프론트에 유효한 토큰이 없다면 인스타 연동 페이지로 이동
 @app.get("/api/v1/authorize")
 def go_to_authorize_page():
-    return RedirectResponse(url=authorize_url)
+    return RedirectResponse(url=insta.authorize_url)
 
 
 # 앱 접속 시 프론트에 토큰이 있다면 리프레쉬 토큰 발급
@@ -105,7 +89,7 @@ def go_to_authorize_page():
 # 프론트에서 발급 받은 토큰 저장 후 user_info_change_by_access_token 호출
 @app.get("/api/v1/refresh_token")
 def get_refresh_token(long_access_token: str):
-    res = requests.get(refresh_token_url+long_access_token).json()
+    res = requests.get(insta.refresh_token_url+long_access_token).json()
     # 토큰이 만료되었다면
     if res['expires_in'] < 30:
         go_to_authorize_page()
@@ -119,7 +103,7 @@ def get_refresh_token(long_access_token: str):
 @app.post("/api/v1/users/by_access_token")
 def user_info_change_by_access_token(access_token: str, db: Session = Depends(get_db)):
     # 엑세스 토큰으로 user 정보 가져옴
-    user_info = get_user_info(access_token=access_token)
+    user_info = insta.get_user_info(access_token=access_token)
 
     # user 업데이트 시도, 성공시 user반환, 실패시 insta_id_not_found 반환
     res = update_user(user=user_info, db=db)
@@ -143,74 +127,10 @@ def get_insta_code(code = None, error = None, error_description = None):
     if code is None:
         raise HTTPException(status_code=404, detail="code is not found")
     # 단기 실행 토큰 발급
-    short_token = get_short_token(code)
+    short_token = insta.get_short_token(code)
     # 장기 실행 토큰 발급
     # {access_token: 'access_token', token_type: 'token_type', expires_in: 5184000}
-    return get_long_token(short_token)
-
-
-# 단기 토큰 얻기
-def get_short_token(code: str):
-    data = {
-        'client_id': app_id,
-        'client_secret': secret_id,
-        'code': code,
-        'grant_type': 'authorization_code',
-        'redirect_uri': redirect_url
-    }
-    headers = {'Content-Type': 'application/x-www-form-urlencoded', 'charset': 'UTF-8', 'Accept': '*/*'}
-    try:
-        res = requests.post(get_short_token_url, headers = headers, data = data).json()
-        # 유효하지 않은 코드면
-        if 'error_type' in res:
-            raise HTTPException(status_code=404, detail=res['error_message'])
-            # 적절한 페이지로 이동시키기
-        else: return res["access_token"]
-    except Exception as ex:
-        print(ex.args)
-
-
-# 장기 토큰 얻기
-def get_long_token(short_access_token: str):
-    try:
-        res = requests.get(get_long_token_url+short_access_token).json()
-        return res
-    except Exception as ex:
-        print(ex.args)
-
-
-# 엑세스 토큰으로 user info 반환
-def get_user_info(access_token: str):
-    try:
-        # 엑세스 토큰을 통해 user 정보 받아오기에 필요한 username을 가져온다.
-        username = requests.get(get_user_name_url + access_token).json()["username"]
-        
-        # 만약 헤더 오류 시 아래 api로 대체
-        # user_info = requests.get(f'https://www.instagram.com/web/search/topsearch/?query={username}')
-        
-        # 헤더 정보에 대해서는 좀 더 알아보고 나중에 수정
-        headers = {
-            'user-agent':'Mozilla/5.0 (iPhone; CPU iPhone OS 12_3_1 like Mac OS X) \
-                AppleWebKit/605.1.15 (KHTML, like Gecko) Mobile/15E148 Instagram 105.0.0.11.118 \
-                    (iPhone11,8; iOS 12_3_1; en_US; en-US; scale=2.00; 828x1792; 165586599)'
-        }
-        # username으로 user 정보 가져오는 api 호출
-        user_info = requests.get(get_user_info_url + username, headers=headers).json()['data']['user']
-        
-        insta_id = user_info['id']
-        full_name = user_info['full_name']
-        follower = user_info['edge_followed_by']['count']
-        following = user_info['edge_follow']['count']
-        profile_image_url = user_info['profile_pic_url']
-
-        user_create_data = schemas.UserCreate(insta_id=insta_id, username=username,\
-        full_name=full_name, follower=follower, following=following, \
-            profile_image_url=profile_image_url)
-
-        return user_create_data
-
-    except Exception as ex:
-        print(ex.args)
+    return insta.get_long_token(short_token)
 
 
 # user 생성에 필요한 정보를 중복 여부에 따라 생성 혹은 업데이트 API 호출
@@ -220,7 +140,7 @@ def create_user(user: schemas.UserCreate, db: Session = Depends(get_db)):
 
 
 # user 업데이트에 필요한 정보를 보내면 user 정보 업데이트
-@app.patch('/api/v1/users', response_model=schemas.User)
+@app.put('/api/v1/users', response_model=schemas.User)
 def update_user(user: schemas.UserCreate, db: Session = Depends(get_db)):
     return crud.update_user(db, user=user)
 
@@ -232,11 +152,10 @@ def show_user(user_id: int, db: Session = Depends(get_db)):
 
 
 # user_id를 path variable로 받아서 해당 유저를 soft delete
-@app.patch('/api/v1/users/{user_id}', response_model=schemas.User)
+@app.put('/api/v1/users/{user_id}', response_model=schemas.User)
 def delete_user(user_id: int, db: Session = Depends(get_db)):
     return crud.delete_user(db=db, user_id=user_id)
 
-# ----------------------------------------------------
 
 # D-6
 # user_id를 path variable로 받아서 user에 해당하는 질문들을 반환
